pyproject.toml
[project]
name = "vision-desktop-automation"
version = "1.0.0"
description = "Vision-based desktop automation with dynamic icon grounding using OCR"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyautogui>=0.9.54",
    "pillow>=10.0.0",
    "mss>=9.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pyperclip>=1.8.0",
    "easyocr>=1.7.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
